telegram_bot_py/requirements.txt
paramiko~=3.4.0
python-dateutil~=2.8.2
python-dotenv~=1.0.0
python-telegram-bot~=20.7
celery[redis]~=5.3.6
redis~=5.0.1
